+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.0.20] - 12.05.2025
+
+### Fixed
+- Issue with PATH - (h/t: @SethTenenbaum)
+- Overloaded functions in mock (functions with same name)
+- Memory issue because of output folder watchers
+- Better Echidna repro (using shrunken version)
+- Refresh contracts button
+
 ## [0.0.19] - 17.04.2025
 
 ### Fixed
@@ -11,5 +20,6 @@
 - Coverage report compatibility with new Medusa report format
 - Issue with fuzzer not stopping
 
+[0.0.20]: https://github.com/Recon-Fuzz/recon-extension/releases/tag/v0.0.20
 [0.0.19]: https://github.com/Recon-Fuzz/recon-extension/releases/tag/v0.0.19
 [0.0.18]: https://github.com/Recon-Fuzz/recon-extension/releases/tag/v0.0.18